requirements.txt
facenet_pytorch==2.6.0
Flask==3.1.1
numpy==2.3.1
Pillow==11.3.0
opencv-python==4.7.0.72
